renamed Configured to Guarded and mirror for fields
