+++ reflection-rewriter/rewriter/src/main/java/io/papermc/reflectionrewriter/DefineClassRule.java
@@ -3,8 +3,8 @@ package io.papermc.reflectionrewriter;
 import io.papermc.asm.ClassInfo;
 import io.papermc.asm.ClassInfoProvider;
 import io.papermc.asm.ClassProcessingContext;
-import io.papermc.asm.rules.MethodRewriteRule;
 import io.papermc.asm.rules.RewriteRule;
+import io.papermc.asm.rules.method.MethodRewriteRule;
 import java.lang.constant.MethodTypeDesc;
 import java.util.Set;
 import org.checkerframework.checker.nullness.qual.NonNull;
