+++ src/main/java/io/papermc/asm/rules/method/MethodRewrites.java
@@ -1,4 +1,4 @@
-package io.papermc.asm.rules;
+package io.papermc.asm.rules.method;
 
 import io.papermc.asm.ClassProcessingContext;
 import io.papermc.asm.rules.builder.matcher.MethodMatcher;
@@ -17,7 +17,7 @@ public final class MethodRewrites {
 
     // Changes a parameter type to a super type. This isn't a compile break, but is an ABI break. We just change the
     // offending parameter in the descriptor and move on.
-    public record SuperTypeParam(Set<Class<?>> owners, MethodMatcher methodMatcher, ClassDesc oldParamType, ClassDesc newParamType) implements MethodRewriteRule {
+    public record SuperTypeParam(Set<Class<?>> owners, MethodMatcher methodMatcher, ClassDesc oldParamType, ClassDesc newParamType) implements GuardedMethodRewriteRule {
 
         @Override
         public Rewrite rewrite(final ClassProcessingContext context, final boolean invokeDynamic, final int opcode, final String owner, final String name, final MethodTypeDesc descriptor, final boolean isInterface) {
@@ -31,7 +31,7 @@ public final class MethodRewrites {
 
     // Changes a return type to a subtype of the old type. This isn't a compile break as subtypes inherit everything, but it is an ABI break.
     // We just change the return type in the descriptor and move on.
-    public record SubTypeReturn(Set<Class<?>> owners, MethodMatcher methodMatcher, ClassDesc oldReturnType, ClassDesc newReturnType) implements MethodRewriteRule {
+    public record SubTypeReturn(Set<Class<?>> owners, MethodMatcher methodMatcher, ClassDesc oldReturnType, ClassDesc newReturnType) implements GuardedMethodRewriteRule {
 
         @Override
         public @Nullable Rewrite rewrite(final ClassProcessingContext context, final boolean invokeDynamic, final int opcode, final String owner, final String name, final MethodTypeDesc descriptor, final boolean isInterface) {
